study_planner/requirements.txt
streamlit
numpy
pandas
python-constraint
matplotlib
plotly
